hpn #014 move tests for Eshva.DockerCompose into subfolders
