+++ sources/tests/Eshva.DockerCompose.Tests.Unit/DownProjectCommand/GivenDownProjectCommandWhenExecuteCalled.cs
@@ -3,7 +3,6 @@
 using System;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
-using Eshva.DockerCompose.Commands.UpProject;
 using Eshva.DockerCompose.Infrastructure;
 using Moq;
 using Xunit;
@@ -11,22 +10,19 @@ using Xunit;
 #endregion
 
 
-namespace Eshva.DockerCompose.Tests.Unit
+namespace Eshva.DockerCompose.Tests.Unit.DownProjectCommand
 {
-    public sealed class GivenUpProjectCommandWhenExecuteCalled
+    public sealed class GivenDownProjectCommandWhenExecuteCalled
     {
         [Fact]
-        public async Task ShouldContainValidArgumentsIfDetached()
+        public async Task ShouldContainValidArguments()
         {
             var starterMock = new Mock<IProcessStarter>();
             Expression<Func<string, bool>> argumentsValidator =
-                arguments => arguments.Equals("-f \"project.yaml\" up --detach", StringComparison.OrdinalIgnoreCase);
+                arguments => arguments.Equals("-f \"project.yaml\" down", StringComparison.OrdinalIgnoreCase);
             starterMock.Setup(starter => starter.Start(It.Is(argumentsValidator), It.IsAny<TimeSpan>()))
                        .Returns(Task.FromResult(0));
-            var command = new UpProjectCommand(starterMock.Object, "project.yaml")
-                          {
-                              IsDetached = true
-                          };
+            var command = new Commands.DownProject.DownProjectCommand(starterMock.Object, "project.yaml");
             await command.Execute();
             starterMock.Verify(
                 starter => starter.Start(It.Is(argumentsValidator), It.IsAny<TimeSpan>()),
