+++ sources/tests/Eshva.DockerCompose.Tests.Unit/ExecutableProcessStarter/GivenExecutableProcessStarterWhenStartCalled.cs
@@ -3,21 +3,20 @@
 using System;
 using System.Threading.Tasks;
 using Eshva.DockerCompose.Exceptions;
-using Eshva.DockerCompose.Infrastructure;
 using FluentAssertions;
 using Xunit;
 
 #endregion
 
 
-namespace Eshva.DockerCompose.Tests.Unit
+namespace Eshva.DockerCompose.Tests.Unit.ExecutableProcessStarter
 {
     public sealed class GivenExecutableProcessStarterWhenStartCalled
     {
         [Fact]
         public void ShouldThrowIfExecutableNotFound()
         {
-            var starter = new ExecutableProcessStarter("unknown-process-name");
+            var starter = new Infrastructure.ExecutableProcessStarter("unknown-process-name");
             Func<Task<int>> start = () => starter.Start(string.Empty, TimeSpan.FromDays(1));
             start.Should().Throw<ProcessStartException>();
         }
@@ -25,7 +24,7 @@ namespace Eshva.DockerCompose.Tests.Unit
         [Fact]
         public void ShouldBreakExecutionAfterSpecifiedAmountOfTime()
         {
-            var starter = new ExecutableProcessStarter("docker-compose");
+            var starter = new Infrastructure.ExecutableProcessStarter("docker-compose");
             Func<Task> execute = async () => await starter.Start(string.Empty, TimeSpan.FromMilliseconds(1));
             execute.Should().ThrowExactly<TimeoutException>();
         }
@@ -33,7 +32,7 @@ namespace Eshva.DockerCompose.Tests.Unit
         [Fact]
         public async Task ShouldCaptureProcessStandardOutput()
         {
-            var starter = new ExecutableProcessStarter("docker");
+            var starter = new Infrastructure.ExecutableProcessStarter("docker");
             await starter.Start("--version", TimeSpan.FromDays(1));
             starter.StandardOutput.ReadToEnd().Length.Should().BeGreaterThan(0);
         }
@@ -41,7 +40,7 @@ namespace Eshva.DockerCompose.Tests.Unit
         [Fact]
         public async Task ShouldCaptureProcessStandardError()
         {
-            var starter = new ExecutableProcessStarter("docker");
+            var starter = new Infrastructure.ExecutableProcessStarter("docker");
             await starter.Start(string.Empty, TimeSpan.FromDays(1));
             starter.StandardError.ReadToEnd().Length.Should().BeGreaterThan(0);
         }
